sage/pyproject.toml
[project]
name = "sage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.7",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "opik>=1.10.5",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "surrealdb>=1.0.8",
    "uvicorn>=0.40.0",
]
